Moved Utility methods in Query to Util
